requirements.txt
aiogram==3.3.0
aiohttp==3.9.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
loguru==0.7.2
